commit 21:20 2024 11 22

+++ src/ft_type_identifier.c
@@ -6,7 +6,7 @@
 /*   By: hoskim <hoskim@student.42prague.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 21:22:14 by hoskim            #+#    #+#             */
-/*   Updated: 2024/10/21 22:05:49 by hoskim           ###   ########.fr       */
+/*   Updated: 2024/10/22 20:58:36 by hoskim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	ft_type_identifier(const char c, va_list *list)
 	else if (c == 'x' || c == 'X')
 		return (ft_hex_printer(va_arg(*list, unsigned int), c));
 	else if (c == 'p')
-		return (ft_ptr_printer(va_arg(*list, unsigned long)));
+		return (ft_ptr_printer(va_arg(*list, unsigned long long)));
 	else if (c == '%')
 		return (ft_char_printer('%'));
 	return (0);
